Refactored sub_slides to no longer connect
