+++ modules/main_Presenter/client/components/PresenterActions.jsx
@@ -1,4 +1,6 @@
 const ADD_PREVIEWS = 'ADD_PREVIEWS';
+const NEXT_SLIDE = 'NEXT_SLIDE';
+const PREV_SLIDE = 'PREV_SLIDE';
 
 export function addPreviews (previews) {
   return {
@@ -7,21 +9,15 @@ export function addPreviews (previews) {
   }
 }
 
-export function getPreviews () {
-  return (dispatch, getState) => {
-    GoogleApi.get('drive/v2/files?q=mimeType="application/vnd.google-apps.presentation"', {}, function (err, result) {
-      if (err) console.error(err);
-      // Map an array of preview objects with specific properties
-      var previews = result.items.map((doc) => {
-        return {
-          link: doc.embedLink.replace('link', 'embed'),
-          title: doc.title,
-          thumbnail: doc.thumbnailLink,
-          gid: doc.id
-        };
-      });
-      console.log(previews);
-      dispatch(addPreviews(previews))
-    })
+export function nextSlide() {
+  return { 
+    type: NEXT_SLIDE
   }
 }
+
+export function prevSlide() { 
+  return { 
+    type: PREV_SLIDE
+  }
+}
+
