--- modules/main_Presenter/client/components/PresenterReducers.jsx
@@ -1,8 +1,0 @@
-export default function counter(state = [], action) {
-  switch (action.type) {
-  case 'ADD_PREVIEWS':
-    return Object.assign([], state, action.payload);
-  default:
-    return state;
-  }
-}
