+++ modules/main_Presenter/client/components/Reducers.jsx
@@ -1,7 +1,15 @@
-export default function counter(state = [], action) {
+import {Map, List} from 'immutable'
+
+const initial = Map({
+  presentation: Map({gid: undefined, index: 0})
+})
+
+export default function counter(state = initial, action) {
   switch (action.type) {
-  case 'ADD_PREVIEWS':
-    return Object.assign([], state, action.payload);
+  case 'NEXT_SLIDE':
+    return state.updateIn(['presentation', 'index'], index => index + 1)
+  case 'PREV_SLIDE':
+    return state.updateIn(['presentation', 'index'], index => index - 1)
   default:
     return state;
   }
