+++ modules/main_Presenter/client/reducers.jsx
@@ -1,8 +1,7 @@
 import { combineReducers } from 'redux'
 
-import addPreviews from './components/PresenterReducers.jsx'
-
+import addPreviews from './components/Reducers.jsx'
 // add reducers here
 export default combineReducers({
-  list: addPreviews,
+  list: addPreviews
 })
