--- modules/sub_Slides/client/Slides.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-
-export default class Slides extends Component {
-  render() {
-    return (
-      < div >
-        hello slide
-      </div>
-    );
-  }
-}
